[APP-237] GA 이벤트 추가 (#260)
# Key Changes

- GA에서 이벤트와 화면 트래킹이 가능하도록 기능을 추가했습니다.

# Details

- AnalyticsService에서 event log를 위한 메서드를 작성하고, 이에 대한 타입을
config/analytics.ts 파일에 정의했습니다.
- 기본 log event name은 아래와 같습니다.
```typescript
export const DEFAULT_LOG_EVENT_NAME = 'unknown';
``` 
- 화면 트래킹 로직이 NaviagtionContainer에 들어가야하기 때문에 해당 로직을 포함하여
CustomNavigationContainer로 추상화했습니다.
- 초기 앱 로딩이 완료된 상태에서 화면을 트래킹 하기 위해서 initLoadingFinishAtom을 작성하고 사용했습니다.
- firebase의 기본 screen tracking을 비활성화했습니다.

# Closes Issue

close #259

+++ src/services/user.ts
@@ -5,6 +5,7 @@ import DeviceService from './device';
 import storeToken from '../utils/storeToken';
 import customShowToast from '../configs/toast';
 import {DeleteUserInfoType, SetUserInfoType} from '../hooks/useUserState';
+import AnalyticsService from './analytics';
 
 type OnRegisterParamsType = {
   accessToken: string;
@@ -12,6 +13,10 @@ type OnRegisterParamsType = {
   setUserInfo: SetUserInfoType;
   /** 바로 로그인하지 않을 때 사용하는 params입니다. */
   setNotLoggedIn?: true;
+};
+type InitializeUserParamsType = {
+  deleteUserInfo: DeleteUserInfoType;
+  user: UserInfoType;
 };
 
 export default class UserService {
@@ -53,26 +58,46 @@ export default class UserService {
 
   // 로그아웃 및 회원 탈퇴
 
-  static async logout(deleteUserInfo: DeleteUserInfoType): Promise<void> {
-    this.initializeUserState(deleteUserInfo);
-    customShowToast('logout');
+  static async logout({
+    deleteUserInfo,
+    user,
+  }: Partial<InitializeUserParamsType>): Promise<void> {
+    if (!deleteUserInfo || !user) {
+      this.initializeUserState(undefined);
+      return;
+    }
+    await AnalyticsService.logAnalyticsEvent('logout', {
+      logoutUserId: user.id,
+    }).finally(() => {
+      this.initializeUserState(deleteUserInfo);
+      customShowToast('logout');
+    });
   }
 
-  static async unregister(deleteUserInfo: DeleteUserInfoType): Promise<void> {
+  static async unregister({
+    deleteUserInfo,
+    user,
+  }: InitializeUserParamsType): Promise<void> {
     try {
       await CoreAPI.unregister();
-      this.initializeUserState(deleteUserInfo);
-      customShowToast('unregister');
+      await AnalyticsService.logAnalyticsEvent('unregister', {
+        unregisterUserId: user.id,
+      }).finally(() => {
+        this.initializeUserState(deleteUserInfo);
+        customShowToast('unregister');
+      });
     } catch (error) {
       customShowToast('unregisterError');
     }
   }
 
-  static initializeUserState(deleteUserInfo: DeleteUserInfoType): void {
+  static initializeUserState(
+    deleteUserInfo: DeleteUserInfoType | undefined,
+  ): void {
     storage.delete('accessToken');
     storage.delete('refreshToken');
     storage.delete('tempToken');
-    deleteUserInfo();
+    if (deleteUserInfo) deleteUserInfo();
     storage.set('isLoggedIn', false);
   }
 }
